Function renaming and review changes implemented
