--- src/functions/on-call/load-on-call-data/mockData.json
@@ -1,7 +1,0 @@
-{
-    "headers": {
-      "Content-Type": "application/json"
-    },
-    "body": "{\"name\": \"Frederic\"}",
-    "queryStringParameters": { "year": "2023" }
-}
