--- src/functions/on-call/weekly-check/mockData.json
@@ -1,6 +1,0 @@
-{
-    "headers": {
-      "Content-Type": "application/json"
-    },
-    "body": "{\"name\": \"Frederic\"}"
-}
